# docs/swagger.yaml
basePath: /service/api
definitions:
...
  dto.BaseResponse:
    properties:
      data: {}
      message_en:
        example: Success
        type: string
      message_th:
        example: สำเร็จ
        type: string
      status:
        example: success
        type: string
      status_code:
        example: 200
        type: integer
    type: object
...
  dto.CreateReceivableDTO:
    properties:
      amount:
        type: number
      balance:
        type: number
      bank_id:
        type: string
      customer:
        type: string
      due_date:
        type: string
      invoice_no:
        type: string
      issue_date:
        type: string
    type: object
...
  dto.RecordReceiptDTO:
    properties:
      amount:
        description: จำนวนเงินที่รับชำระ
        type: number
      bank_id:
        description: รหัสบัญชีธนาคารที่รับชำระ
        type: string
      note:
        description: หมายเหตุเพิ่มเติม
        type: string
      payment_date:
        description: วันที่รับชำระ (รูปแบบ YYYY-MM-DD)
        type: string
      payment_method:
        description: วิธีชำระเงิน เช่น เงินสด โอน บัตร
        type: string
      payment_ref:
        description: เลขอ้างอิงการชำระ เช่น สลิป/เลขที่ธุรกรรม
        type: string
      receivable_id:
        description: รหัสรายการลูกหนี้ที่รับชำระ
        type: string
    type: object
...
  dto.UpdateReceivableDTO:
    properties:
      amount:
        type: number
      balance:
        type: number
      bank_id:
        type: string
      customer:
        type: string
      due_date:
        type: string
      invoice_no:
        type: string
      issue_date:
        type: string
      status:
        type: string
    type: object
...
host: api.dev.rkp-media.com
info:
  contact: {}
  description: This is an ERP API demo.
  title: ERP Demo API
  version: "1.0"
paths:
...
  /v1/receivable/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a receivable record by its ID
      parameters:
      - description: Receivable ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Delete receivable by ID
      tags:
      - Receivables
    get:
      consumes:
      - application/json
      description: Get a receivable record by its ID
      parameters:
      - description: Receivable ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get receivable by ID
      tags:
      - Receivables
    put:
      consumes:
      - application/json
      description: Update a receivable record by its ID
      parameters:
      - description: Receivable ID
        in: path
        name: id
        required: true
        type: string
      - description: Receivable data to update
        in: body
        name: receivable
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReceivableDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Update receivable by ID
      tags:
      - Receivables
  /v1/receivable/create:
    post:
      consumes:
      - application/json
      description: Create a new receivable record
      parameters:
      - description: Receivable data
        in: body
        name: receivable
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReceivableDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Create a new receivable
      tags:
      - Receivables
  /v1/receivable/list:
    get:
      consumes:
      - application/json
      description: Get a paginated list of receivables
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort by field
        enum:
        - created_at
        - updated_at
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List receivables
      tags:
      - Receivables
  /v1/receivable/record-receipt:
    post:
      consumes:
      - application/json
      description: Record a receipt for a receivable
      parameters:
      - description: Receipt data
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/dto.RecordReceiptDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Record receipt
      tags:
      - Receivables
  /v1/receivable/summary:
    get:
      consumes:
      - application/json
      description: Get summary of receivables
      parameters:
      - description: Bank ID
        in: query
        name: bank_id
        type: string
      - description: Report type
        enum:
        - day
        - month
        - all
        in: query
        name: report
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Summary receivables
      tags:
      - Receivables
...
swagger: "2.0"